Suspected cause of current problem:  not referencing correctly the functions (hpp file had functions with {}, which told Luigi to  operate on an empty method).
 Possible solution:
Refactored SumoMvmt.hpp and SumoMvmt.cpp to properly call SumoMvmt.cpp by removing the empty function bodies in the hpp file, declaring the pins for the motor driver in hpp file and then defining them in cpp file (multiple definition exception was thrown). If this doesn't work then the content of SumoMvmt.cpp will be moved to the corresponding headerfile. TO BE TESTED